All the changes in wfa_cmdproc.c, wfa_cs.c, and wfa_tg.c are related to SIG-523, i.e., the framerate is declared as float type instead of int type.
SCR#SIG-523
close[s:jira, t:523]

+++ lib/wfa_cmdproc.c
@@ -750,8 +750,13 @@ int xcCmdProcAgentSendPing(char *pcmdStr, BYTE *aBuf, int *aLen)
         if(strcasecmp(str, "frameRate") == 0)
         {
             str = strtok_r(NULL, ",", &pcmdStr);  
-            staping->frameRate=atoi(str);
-            DPRINT_INFO(WFA_OUT, "framerate %i\n", staping->frameRate);
+            //<<<------------------------------------------------------------------------
+            //jira issue: SIG-523
+            staping->frameRate=atof(str);
+            DPRINT_INFO(WFA_OUT, "framerate %f\n", staping->frameRate);
+            ////staping->frameRate=atoi(str);
+            ////DPRINT_INFO(WFA_OUT, "framerate %i\n", staping->frameRate);
+            //------------------------------------------------------------------------>>>
         }
         if(strcasecmp(str, "duration") == 0)
         {
